extract audio process to feature extractor
